Pull the form part of privacy declaration into its own component
