+++ clients/ctl/admin-ui/src/features/system/PrivacyDeclarationStep.tsx
@@ -5,34 +5,12 @@ import { Form, Formik, FormikHelpers } from "formik";
 import React, { Fragment, useState } from "react";
 import * as Yup from "yup";
 
-import { useAppSelector } from "~/app/hooks";
-import {
-  CustomMultiSelect,
-  CustomSelect,
-  CustomTextInput,
-} from "~/features/common/form/inputs";
 import { getErrorMessage, isErrorResult } from "~/features/common/helpers";
 import { AddIcon } from "~/features/common/Icon";
-import {
-  selectDataQualifiers,
-  useGetAllDataQualifiersQuery,
-} from "~/features/data-qualifier/data-qualifier.slice";
-import {
-  selectDataSubjects,
-  useGetAllDataSubjectsQuery,
-} from "~/features/data-subjects/data-subject.slice";
-import {
-  selectDataUses,
-  useGetAllDataUsesQuery,
-} from "~/features/data-use/data-use.slice";
-import {
-  selectDataCategories,
-  useGetAllDataCategoriesQuery,
-} from "~/features/taxonomy/taxonomy.slice";
 import { PrivacyDeclaration, System } from "~/types/api";
 
 import PrivacyDeclarationAccordion from "./PrivacyDeclarationAccordion";
-import PrivacyDeclarationFormExtension from "./PrivacyDeclarationFormExtension";
+import PrivacyDeclarationForm from "./PrivacyDeclarationForm";
 import { useUpdateSystemMutation } from "./system.slice";
 
 type FormValues = PrivacyDeclaration;
@@ -75,17 +53,6 @@ const PrivacyDeclarationStep = ({
   >(system?.privacy_declarations ? [...system.privacy_declarations] : []);
   const [updateSystem] = useUpdateSystemMutation();
   const [isLoading, setIsLoading] = useState(false);
-
-  // Query subscriptions:
-  useGetAllDataCategoriesQuery();
-  useGetAllDataSubjectsQuery();
-  useGetAllDataQualifiersQuery();
-  useGetAllDataUsesQuery();
-
-  const allDataCategories = useAppSelector(selectDataCategories);
-  const allDataSubjects = useAppSelector(selectDataSubjects);
-  const allDataUses = useAppSelector(selectDataUses);
-  const allDataQualifiers = useAppSelector(selectDataQualifiers);
 
   const initialValues: PrivacyDeclaration = {
     name: "",
@@ -188,52 +155,7 @@ const PrivacyDeclarationStep = ({
                 <Divider m="0px !important" />
               </Fragment>
             ))}
-            <Stack spacing={4}>
-              <CustomTextInput
-                name="name"
-                label="Declaration name"
-                tooltip="A system may have multiple privacy declarations, so each declaration should have a name to distinguish them clearly."
-              />
-              <CustomMultiSelect
-                name="data_categories"
-                label="Data categories"
-                options={allDataCategories?.map((data) => ({
-                  value: data.fides_key,
-                  label: data.fides_key,
-                }))}
-                tooltip="What type of data is your system processing? This could be various types of user or system data."
-              />
-              <CustomSelect
-                id="data_use"
-                label="Data use"
-                name="data_use"
-                options={allDataUses.map((data) => ({
-                  value: data.fides_key,
-                  label: data.fides_key,
-                }))}
-                tooltip="What is the system using the data for. For example, is it for third party advertising or perhaps simply providing system operations."
-              />
-              <CustomMultiSelect
-                name="data_subjects"
-                label="Data subjects"
-                options={allDataSubjects.map((data) => ({
-                  value: data.fides_key,
-                  label: data.fides_key,
-                }))}
-                tooltip="Whose data are you processing? This could be customers, employees or any other type of user in your system."
-              />
-              <CustomSelect
-                id="data_qualifier"
-                label="Data qualifier"
-                name="data_qualifier"
-                options={allDataQualifiers.map((data) => ({
-                  value: data.fides_key,
-                  label: data.fides_key,
-                }))}
-                tooltip="How identifiable is the user in the data in this system? For instance, is it anonymized data where the user is truly unknown/unidentifiable, or it is partially identifiable data?"
-              />
-              {!abridged ? <PrivacyDeclarationFormExtension /> : null}
-            </Stack>
+            <PrivacyDeclarationForm abridged={abridged} />
             <Box>
               <Button
                 type="submit"
